[ADD] 게시글 댓글 카운트 조회 기능 추가

+++ service/comment/src/main/java/kuke/board/comment/entity/Comment.java
@@ -41,7 +41,7 @@ public class Comment {
     }
 
     public boolean isRoot() {
-        return parentCommentId == commentId;
+        return parentCommentId.longValue() == commentId.longValue();
     }
 
     public void delete() {
